removed Authentication Controller, added success page, fixed most issues with login/ still need to fix admin default? still need more security and for validation.
