+++ src/main/java/com/pqtca/demo/SecurityConfig.java
@@ -36,13 +36,13 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
         http
                 /* Login configuration */
                 .formLogin()
-                .loginPage("/adminLogin")
-                .defaultSuccessUrl("/") // user's home page, it can be any URL
+                .loginPage("/admin/login")
+                .defaultSuccessUrl("/admin/success") // user's home page, it can be any URL
                 .permitAll() // Anyone can go to the login page
                 /* Logout configuration */
                 .and()
                 .logout()
-                .logoutSuccessUrl("/login?logout") // append a query string value
+                .logoutSuccessUrl("/admin/login?logout") // append a query string value
                 /* Pages that can be viewed without having to log in */
                 .and()
                 .authorizeRequests()
