--- src/main/java/com/pqtca/demo/controllers/AuthenticationController.java
@@ -1,51 +1,0 @@
-package com.pqtca.demo.controllers;
-
-
-import com.pqtca.demo.models.User;
-import com.pqtca.demo.repos.UserRepo;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
-
-
-
-@Controller
-public class AuthenticationController {
-
-    private UserRepo usersDao;
-    private PasswordEncoder passwordEncoder;
-
-    public AuthenticationController(UserRepo usersDao, PasswordEncoder passwordEncoder) {
-        this.usersDao = usersDao;
-        this.passwordEncoder = passwordEncoder;
-    }
-
-    @GetMapping("/admin/register")
-    public String showSignUpForm(Model model) {
-        User user = new User();
-        model.addAttribute("user", user);
-        return "admin/register";
-    }
-
-    @PostMapping("/admin/register")
-    public String registerUser(@ModelAttribute User user) {
-        System.out.println("here is the user email " + user.getEmail());
-        System.out.println("here is the user " + user.getUsername());
-        String hash = passwordEncoder.encode(user.getPassword());
-        user.setPassword(hash);
-        usersDao.save(user);
-        return "redirect:/admin/login";
-    }
-
-    @GetMapping("/admin/login")
-    public String showLoginForm() {
-        User loggedInUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        return "/admin/login";
-    }
-
-
-}
