+++ src/main/resources/templates/admin/login.html
@@ -1,14 +1,25 @@
 <!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
+<html xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.w3.org/1999/xhtml">
 <head>
 <!--<head th:replace="partials/head :: head('Login')">-->
     <meta charset="UTF-8" />
     <title>Log In</title>
+    <div sec:authorize="isAuthenticated()">
+        <!--This content is only shown to authenticated users.-->
+
+        <!--User ID: <span th:text="${#authentication.principal.id}"/>-->
+        <!--Username: <span th:text="${#authentication.principal.username}"/>-->
+        <!--Email: <span th:text="${#authentication.principal.email}"/>-->
+    </div>
+
 </head>
 <body>
 <!--<div th:replace="partials/navbar :: navbar"></div>-->
+<a href="/admin/register">create new admin</a>
+
 
 <h1>Log In</h1>
+
 <div th:if="${param.error}">
     Invalid Username and Password.
 </div>
@@ -16,7 +27,7 @@
     You have been logged out.
 </div>
 
-<form th:action="@{/login}" method="post">
+<form th:action="@{/admin/login}" th:method="post">
     <label for="username">Username</label>
     <input id="username" name="username" type="text" />
     <label for="password">Password</label>
