+++ src/main/resources/templates/admin/register.html
@@ -1,12 +1,18 @@
 <!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-<!--<head th:replace="partials/head :: head('Register')">-->
-    <meta charset="UTF-8" />
-    <title>admin Register</title>
+<html xmlns:th="http://www.thymeleaf.org/"
+      xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity4/">>
+<head th:replace="fragments/head :: head('Admin Register')">
+
 </head>
 <body>
 <!--<div th:replace="partials/navbar :: navbar"></div>-->
+<div sec:authorize="isAuthenticated()">
+    <!--This content is only shown to authenticated users.-->
+
+    <!--User ID: <span th:text="${#authentication.principal.id}"/>-->
+    Username: <span th:text="${#authentication.principal.username}"/>
+    <!--Email: <span th:text="${#authentication.principal.email}"/>-->
+</div>
 
 <h1>Create an account</h1>
 <form th:action="@{/admin/register}" th:method="POST" th:object="${user}">
